Play with superclass
